Completely rid the codebase of position handling
It is simply impossible to work with the current implementation properly
in a terraform context because a change to one resource's position would
alter the position of everything else above or below it.

This commit completely removes all code and mention of `position` field.

A more proper and more terraform-friendly approach would be to dedicate
a new specific resource to the task of ordering things itself so that a
change to one item's position in the list do not cascade to changes to
everything else around it, only the ordering resource.